Move base Plotter and related functions out of __init__ of the vis package into separate module parameters
