+++ masci_tools/vis/bokeh_plotter.py
@@ -11,10 +11,10 @@
 #                                                                             #
 ###############################################################################
 """
-Here the :py:class:`masci_tools.vis.Plotter` subclass for the bokeh plotting backend
+Here the :py:class:`masci_tools.vis.parameters.Plotter` subclass for the bokeh plotting backend
 is defined with default values and many helper methods
 """
-from masci_tools.vis import Plotter, _generate_plot_parameters_table
+from .parameters import Plotter, _generate_plot_parameters_table
 import copy
 
 
@@ -26,7 +26,7 @@ class BokehPlotter(Plotter):
     Kwargs in the __init__ method are forwarded to setting default values for the instance
 
     For specific documentation about the parameter/defaults handling refer to
-    :py:class:`~masci_tools.vis.Plotter`.
+    :py:class:`~masci_tools.vis.parameters.Plotter`.
 
     Below the current defined default values are shown:
